Relative path functionality!
autotest can be called from anywhere effectively. Since the files need to be in verif/data anyway and the  Makefile is called in verif/scripts. the data will always be in the same place, no need to add absolute path for data.

(I did not read the readme before trying to use the autograder and lost 2 hours on this relative vs absolute path call)